relocate test file stop versioning manifest production code
