+++ controllers/kyma_controller/suite_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controllers_test
+package kyma_controller_test
 
 import (
 	"context"
