+++ internal/manifest/client.go
@@ -1,4 +1,4 @@
-package v1beta1
+package manifest
 
 import (
 	"context"
