+++ internal/manifest/manifest_controller/ready_check_test.go
@@ -1,4 +1,4 @@
-package v1beta1_test
+package manifest_controller_test
 
 import (
 	"encoding/json"
@@ -6,7 +6,7 @@ import (
 	"path/filepath"
 
 	declarative "github.com/kyma-project/lifecycle-manager/internal/declarative/v2"
-	internalV1beta1 "github.com/kyma-project/lifecycle-manager/internal/manifest/v1beta1"
+	"github.com/kyma-project/lifecycle-manager/internal/manifest"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -33,12 +33,12 @@ var _ = Describe("Custom Manifest consistency check, given Manifest CR with OCI 
 		},
 	)
 	It("Install OCI specs including an nginx deployment", func() {
-		manifest := NewTestManifest("custom-check-oci")
-		manifestName := manifest.GetName()
+		testManifest := NewTestManifest("custom-check-oci")
+		manifestName := testManifest.GetName()
 		validImageSpec := createOCIImageSpec(installName, server.Listener.Addr().String(), false)
 		imageSpecByte, err := json.Marshal(validImageSpec)
 		Expect(err).ToNot(HaveOccurred())
-		Expect(installManifest(manifest, imageSpecByte, false)).To(Succeed())
+		Expect(installManifest(testManifest, imageSpecByte, false)).To(Succeed())
 
 		Eventually(expectManifestStateIn(declarative.StateReady), standardTimeout, standardInterval).
 			WithArguments(manifestName).Should(Succeed())
@@ -55,13 +55,13 @@ var _ = Describe("Custom Manifest consistency check, given Manifest CR with OCI 
 		Expect(resources).ToNot(BeEmpty())
 
 		By("Executing the custom readiness check")
-		customReadyCheck := internalV1beta1.NewManifestCustomResourceReadyCheck()
-		stateInfo, err := customReadyCheck.Run(ctx, testClient, manifest, resources)
+		customReadyCheck := manifest.NewManifestCustomResourceReadyCheck()
+		stateInfo, err := customReadyCheck.Run(ctx, testClient, testManifest, resources)
 		Expect(err).NotTo(HaveOccurred())
 		Expect(stateInfo.State).To(Equal(declarative.StateReady))
 
 		By("cleaning up the manifest")
-		Eventually(deleteManifestAndVerify(manifest), standardTimeout, standardInterval).Should(Succeed())
+		Eventually(deleteManifestAndVerify(testManifest), standardTimeout, standardInterval).Should(Succeed())
 	})
 })
