+++ internal/manifest/spec_resolver.go
@@ -1,4 +1,4 @@
-package v1beta1
+package manifest
 
 import (
 	"context"
